enable scheduler
now, scheduler is enabled, I test it using two functions,
these functions can run alternatively. but I have not found
the solusion to slove the issue of crash after one of the function
terminals. I don't know how to tell the scheduler that infomation
and let the scheduler delete that task.