impliment a little shell for horos
as one of my goals, adding a shell to horos is a
big success. I have add interruption system, memory management,
scheduler and now a shell. I think there is all subsystem except
file system in my horos.
however, even I initially impliment most of the kernel subsystem,
they are not work properly and systematically. so there are lots
of things to do.